model: Add the class `PackageLicenseInfo`
The class is a model for all license information related to a single
package or project. It will be used as input for a new license resolver
class in a later commit.

While all the data in this new model is already available in the
`OrtResult`, adding the model is benefitial, because it simplifies the
implementation of the license resolver mainly for these reasons:

1. The resolver does not have to search through the `OrtResult` to
   gather the information which simplifies the code.
2. The license info can be cached which can speed up subsequent
   requests.
3. The tests of the license resolver only need to provide license
   related data instead of complete `OrtResult`s which contain a lot of
   irrelevant data.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@here.com>